fix(enterprise): close 10 defects found auditing the policy-enforcement wiring
An independent audit of the enterprise policy gates, the PostgreSQL
scaffold and the resource-grant matrix. Every finding was re-verified
against the code before being acted on.

High severity:

* The IP allowlist could only ever DENY in production. resolve_caller_ip
  reads axum's ConnectInfo extension, but both production serve sites
  passed a bare Router -- into_make_service_with_connect_info appeared
  nowhere in the repo except a comment claiming it was already used. So
  every directly-connected request resolved to no IP, and the moment an
  admin enabled the allowlist every direct client was 403'd regardless of
  the list's contents, including the admin's own request to disable it.
  The unit tests "proving" the strict path worked passed only because
  they hand-injected ConnectInfo, a state production never produced.

* A wildcard resource grant leaked across tenants. widen_with_grants
  emitted `OR 1 = 1` for the wildcard case, discarding the entire
  predicate -- including the scope clause that is the ONLY tenant
  dimension these three registries have (they carry no tenant_id column;
  012_collaboration_tenant_scope.sql's header calls the read-side
  equivalent a security bug). Explicit-id grants bypassed it the same
  way, and grant_resource cannot verify resource ownership across a crate
  boundary. Split out member_tenant_reach_where so both branches keep the
  tenant clause: a grant widens visibility, never tenancy.

* `limit as u32` truncation inverted the send-rate limit. The value is an
  unvalidated i64 from the admin console; a deliberately-huge
  "effectively unlimited" number like 2^32 wrapped to 0, and since the
  test is `count >= limit`, `0 >= 0` refused every send for the tenant
  forever. Now saturates via u32::try_from.

Medium:

* PolicyGrace expired 30 minutes after boot on any deployment whose
  traffic is all ungoverned: four reachable-but-not-governed branches
  never called answered(), so a single transient DB error afterwards was
  treated as a dead policy plane and 500'd every authenticated request --
  exactly what those gates' comments forbid.

* A removed member's API key kept working. It has no session generation
  to rotate, so invalidate_user_tokens never touched it. Added
  revoke_api_keys_for_user plus ApiKeyRevoker, and
  CompositeCredentialRevoker because OrgService has one revoker slot and
  there are now two credential kinds that outlive JWT rotation.

* A granted knowledge document was listed but unsearchable: search_rag
  did not apply the widening list_rag_documents does, so the grant was
  half-effective in the one place it changes agent behaviour.
  lexical_candidates now takes acl_binds so it can carry a widened
  multi-placeholder predicate.

* model_channel grants were accepted, stored and reported effective while
  nothing enforced them. list_provider_channels now widens like its
  sibling registries. `employee` grants remain unenforced and are now
  documented as such rather than implied covered.

Low / docs:

* grant_resource validated trimmed ids but stored the raw ones, so " * "
  persisted as a grant matching neither the wildcard sentinel nor any
  real id, rendered as active by the console.

* The Postgres users migration declared session_generation INTEGER while
  every Rust layer models it i64; sqlx's Postgres decode is an exact type
  match, so the first read would fail with ColumnDecode. Now BIGINT,
  verified against a real postgres:16-alpine instance.

* Two doc comments asserted that nothing reads effective_resource_ids and
  the registries gate purely on scope/visibility. Both are now false --
  four live read paths -- and someone reading them would conclude a
  grant-semantics change had zero blast radius.

Verified: dream-domain-devops 79, dream-domain-platform 40,
dream-core-ai-agent and dream-core-app --features enterprise full suites,
dream-core-db including a real-Postgres migration round-trip; both
editions' cargo test --no-run (which caught 7 test-only call sites cargo
check misses). Regression tests added for the cross-tenant boundary, API
key revocation and the trim semantics.

+++ crates/dream-core-db/migrations_postgres/001_users.sql
@@ -32,7 +32,12 @@ CREATE TABLE IF NOT EXISTS users (
     data_secret        TEXT,
     status             TEXT NOT NULL DEFAULT 'active'
                            CHECK (status IN ('active', 'disabled')),
-    session_generation INTEGER NOT NULL DEFAULT 0,
+    -- BIGINT, not INTEGER: every Rust layer models this as `i64`
+    -- (`models/user.rs`, `dream-core-auth::jwt`, `dream-core-api-types::auth`),
+    -- and sqlx's Postgres decode is an exact type match — `i64` against an
+    -- INT4 column fails at read time with a ColumnDecode error even though
+    -- the value always fits.
+    session_generation BIGINT NOT NULL DEFAULT 0,
     created_at         BIGINT NOT NULL,
     updated_at         BIGINT NOT NULL,
     last_login         BIGINT,
